Refactor HubContext
